requirements.txt
asgiref==3.8.1
cachetools==5.5.0
Django==5.1.1
gunicorn==20.1.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httplib2==0.22.0
oauth2client==4.1.3
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.1

